GCP Minecraft Bot/requirements.txt
discord.py>=2.0.0
mcstatus>=9.0.0
